Add files via upload
Create a system which includes a Database, backend and frontend. Chose Python for backend and
your choice of framework and language for frontend and database. Deploy it at your preferred 
hosting solution ideally using free tier, for example, AWS, Azure, Heroku etc. 
All calculations should happen in backend and not in the frontend. All data from the forms should be 
saved in the database and not cached in the frontend. 
If something is ambiguous, please use your best judgement. Use 1-3 hour for this delivery, it is fine 
to take shortcuts to deliver it, please note them down. 
You can use any authentication system, integrated with 3rd party or build your own. There is no need 
to setup CI/CD but a git repo or any other source code management system is important.